+++ packages/medusa-react/.eslintrc.js
@@ -1,12 +1,12 @@
 module.exports = {
-  "extends": [
+  extends: [
     "react-app",
     "prettier/@typescript-eslint",
-    "plugin:prettier/recommended"
+    "plugin:prettier/recommended",
   ],
-  "settings": {
-    "react": {
-      "version": "detect"
-    }
-  }
+  settings: {
+    react: {
+      version: "detect",
+    },
+  },
 }