+++ packages/medusa/src/api/index.js
@@ -38,6 +38,7 @@ export * from "./routes/admin/returns"
 export * from "./routes/admin/tax-rates"
 export * from "./routes/admin/shipping-options"
 export * from "./routes/admin/regions"
+export * from "./routes/admin/product-tags"
 
 // Store
 export * from "./routes/store/auth"