+++ packages/medusa/src/types/orders.ts
@@ -1,12 +1,12 @@
 import { Type } from "class-transformer"
 import {
-  IsString,
-  IsOptional,
   IsArray,
   IsEnum,
-  ValidateNested,
-  IsNotEmpty,
   IsInt,
+  IsNotEmpty,
+  IsOptional,
+  IsString,
+  ValidateNested,
 } from "class-validator"
 import { Order } from "../models/order"
 import { DateComparisonOperator } from "./common"
@@ -66,7 +66,7 @@ export class AdminListOrdersSelector {
   fulfillment_status?: string[]
 
   @IsArray()
-  @IsEnum(PaymentStatus)
+  @IsEnum(PaymentStatus, { each: true })
   @IsOptional()
   payment_status?: string[]